+++ gridopt/power_flow/ac_pf.py
@@ -7,6 +7,7 @@
 #*****************************************************#
 
 from __future__ import print_function
+import time
 import numpy as np
 from .method_error import *
 from .method import PFmethod
@@ -140,7 +141,7 @@ class ACPF(PFmethod):
                 else:
                     assert(net.num_fixed == net.get_num_buses_reg_by_gen()*net.num_periods)
             except AssertionError:
-                raise PFmethodError_BadProblem(self)  
+                raise PFmethodError_BadProblem()  
             
             # Set up problem
             problem = pfnet.Problem(net)
@@ -218,7 +219,7 @@ class ACPF(PFmethod):
                                          net.get_num_tap_changers() +
                                          net.get_num_switched_shunts())*net.num_periods)
             except AssertionError:
-                raise PFmethodError_BadProblem(self)
+                raise PFmethodError_BadProblem()
 
             # Set up problem
             problem = pfnet.Problem(net)
@@ -236,7 +237,7 @@ class ACPF(PFmethod):
         # Invalid
         #########
         else:
-            raise PFmethodError_BadOptSolver(self)
+            raise PFmethodError_BadOptSolver()
             
     def solve(self,net):
 
@@ -262,12 +263,17 @@ class ACPF(PFmethod):
         elif optsolver_name == 'nr':
             optsolver = OptSolverNR()
         else:
-            raise PFmethodError_BadOptSolver(self)
+            raise PFmethodError_BadOptSolver()
         optsolver.set_parameters(optsolver_params[optsolver_name])
 
-        # Problem
-        problem = self.create_problem(net)
+        # Copy network
+        net = net.get_copy()
 
+        # Problem
+        t0 = time.time()
+        problem = self.create_problem(net)
+        problem_time = time.time()-t0
+        
         # Callbacks
         def c1(s):
             if (s.k != 0 and
@@ -306,65 +312,31 @@ class ACPF(PFmethod):
         optsolver.set_info_printer(info_printer)
         
         # Solve
+        t0 = time.time()
         try:
             optsolver.solve(problem)
         except OptSolverError as e:
-            raise PFmethodError_SolverError(self,e)
+            raise PFmethodError_SolverError(e)
         finally:
             
-            # Get results
-            self.set_status(optsolver.get_status())
-            self.set_error_msg(optsolver.get_error_msg())
-            self.set_iterations(optsolver.get_iterations())
-            self.set_primal_variables(optsolver.get_primal_variables())
-            self.set_dual_variables(optsolver.get_dual_variables())
-            self.set_net_properties(net.get_properties())
-            self.set_problem(problem)
-           
-            # Restore net properties
+            # Update network
+            net.set_var_values(optsolver.get_primal_variables()[:net.num_vars])
             net.update_properties()
+            net.clear_sensitivities()
+            if optsolver_name != 'nr':
+                problem.store_sensitivities(*optsolver.get_dual_variables())
+
+            # Save results
+            self.set_solver_status(optsolver.get_status())
+            self.set_solver_message(optsolver.get_error_msg())
+            self.set_solver_iterations(optsolver.get_iterations())
+            self.set_solver_time(time.time()-t0)
+            self.set_solver_primal_variables(optsolver.get_primal_variables())
+            self.set_solver_dual_variables(optsolver.get_dual_variables())
+            self.set_problem(None) # skip for now
+            self.set_problem_time(problem_time)
+            self.set_network_snapshot(net)
  
-    def update_network(self,net):
-
-        # Parameters
-        optsolver_name = self._parameters['optsolver']
-        
-        # Get data
-        problem = self.results['problem']
-        x = self.results['primal variables']
-        lam,nu,mu,pi = self.results['dual variables']
-       
-        # No problem
-        if problem is None:
-            raise PFmethodError_NoProblem(self)
- 
-        # Checks
-        if optsolver_name != 'nr':
-            assert(problem.x.shape == x.shape)
-            assert(net.num_vars+problem.num_extra_vars == x.size)
-            assert(problem.A.shape[0] == lam.size)
-            assert(problem.f.shape[0] == nu.size)
-            assert(problem.G.shape[0] == mu.size)
-            assert(problem.G.shape[0] == pi.size)
-        else:
-            assert(problem.x.shape == x.shape)
-            assert(net.num_vars == x.size)
-            assert(lam is None or not lam.size)
-            assert(nu is None or not nu.size)
-            assert(mu is None or not mu.size)
-            assert(pi is None or not pi.size)
-
-        # Network quantities
-        net.set_var_values(x[:net.num_vars])
-
-        # Network properties
-        net.update_properties()
-        
-        # Network sensitivities
-        net.clear_sensitivities()
-        if optsolver_name != 'nr':
-            problem.store_sensitivities(lam,nu,np.zeros(0),np.zeros(0))
-
     def get_info_printer(self):
 
         # Parameters
@@ -411,7 +383,7 @@ class ACPF(PFmethod):
         # Invalid
         #########
         else:
-            raise PFmethodError_BadOptSolver(self)
+            raise PFmethodError_BadOptSolver()
 
     def apply_shunt_v_regulation(self,solver):